Rename suspensionOfVisits model file.
